broke out placedgate
