+++ client/src/App.tsx
@@ -3,6 +3,7 @@ import './App.css';
 import './circuitboard.css';
 import './toolbar.css';
 import Toolbar from './toolbar';
+import PlacedGate from './placedGate';
 import {DndContext} from '@dnd-kit/core';
 import {useDraggable, useDroppable} from '@dnd-kit/core';
 import {CSS} from '@dnd-kit/utilities';
@@ -214,21 +215,6 @@ function App() {
       } 
     });
   }*/
-}
-
-function PlacedGate(props:any){
-  
-  // Display nothing if there is no placed gate (which is the same as the identity gate).
-  if(props.gateType != "I"){
-    return (
-      <button className = "placedGate">
-        <h1>{props.name}</h1>
-      </button>
-    );
-  } 
-  else return null;
-  
-  
 }
 
 /*function PlacedControlGate(props:any, event:any){
